Refactor models.py enums into core_enums.py.
