+++ api/features_api_test.py
@@ -19,6 +19,7 @@ from unittest import mock
 import werkzeug.exceptions  # Flask HTTP stuff.
 
 from api import features_api
+from internals import core_enums
 from internals import models
 from framework import ramcache
 
@@ -31,7 +32,7 @@ class FeaturesAPITestDelete(testing_config.CustomTestCase):
     self.feature_1 = models.Feature(
         name='feature one', summary='sum', category=1, visibility=1,
         standardization=1, web_dev_views=1, impl_status_chrome=1,
-        intent_stage=models.INTENT_IMPLEMENT)
+        intent_stage=core_enums.INTENT_IMPLEMENT)
     self.feature_1.put()
     self.feature_id = self.feature_1.key.integer_id()
 
@@ -101,7 +102,7 @@ class FeaturesAPITestGet(testing_config.CustomTestCase):
         name='feature one', summary='sum Z',
         owner=['feature_owner@example.com'],
         category=1, visibility=1, standardization=1, web_dev_views=1,
-        impl_status_chrome=5, intent_stage=models.INTENT_IMPLEMENT,
+        impl_status_chrome=5, intent_stage=core_enums.INTENT_IMPLEMENT,
         shipped_milestone=1)
     self.feature_1.put()
     self.feature_1_id = self.feature_1.key.integer_id()
@@ -110,7 +111,7 @@ class FeaturesAPITestGet(testing_config.CustomTestCase):
         name='feature two', summary='sum K',
         owner=['other_owner@example.com'],
         category=1, visibility=1, standardization=1, web_dev_views=1,
-        impl_status_chrome=5, intent_stage=models.INTENT_IMPLEMENT,
+        impl_status_chrome=5, intent_stage=core_enums.INTENT_IMPLEMENT,
         shipped_milestone=2)
     self.feature_2.put()
     self.feature_2_id = self.feature_2.key.integer_id()
@@ -119,7 +120,7 @@ class FeaturesAPITestGet(testing_config.CustomTestCase):
         name='feature three', summary='sum A',
         owner=['other_owner@example.com'],
         category=1, visibility=1, standardization=1, web_dev_views=1,
-        impl_status_chrome=5, intent_stage=models.INTENT_IMPLEMENT,
+        impl_status_chrome=5, intent_stage=core_enums.INTENT_IMPLEMENT,
         shipped_milestone=2, unlisted=True)
     self.feature_3.put()
     self.feature_3_id = self.feature_3.key.integer_id()
