+++ internals/detect_intent_test.py
@@ -18,6 +18,7 @@ import flask
 from unittest import mock
 import werkzeug
 
+from internals import core_enums
 from internals import models
 from internals import approval_defs
 from internals import detect_intent
@@ -31,7 +32,7 @@ class FunctionTest(testing_config.CustomTestCase):
     self.feature_1 = models.Feature(
         name='feature one', summary='detailed sum', category=1, visibility=1,
         standardization=1, web_dev_views=1, impl_status_chrome=1,
-        intent_stage=models.INTENT_IMPLEMENT)
+        intent_stage=core_enums.INTENT_IMPLEMENT)
     self.feature_1.put()
 
   def tearDown(self):
@@ -324,7 +325,7 @@ class IntentEmailHandlerTest(testing_config.CustomTestCase):
     self.feature_1 = models.Feature(
         name='feature one', summary='detailed sum', category=1, visibility=1,
         standardization=1, web_dev_views=1, impl_status_chrome=1,
-        intent_stage=models.INTENT_IMPLEMENT)
+        intent_stage=core_enums.INTENT_IMPLEMENT)
     self.feature_1.put()
     self.feature_id = self.feature_1.key.integer_id()
