+++ internals/processes_test.py
@@ -18,6 +18,7 @@ import testing_config  # Must be imported before the module under test.
 from unittest import mock
 
 from internals import approval_defs
+from internals import core_enums
 from internals import models
 from internals import processes
 
@@ -99,10 +100,10 @@ class HelperFunctionsTest(testing_config.CustomTestCase):
     """A review step is done if the review has completed or was N/a."""
     self.assertFalse(processes.review_is_done(None))
     self.assertFalse(processes.review_is_done(0))
-    self.assertFalse(processes.review_is_done(models.REVIEW_PENDING))
-    self.assertFalse(processes.review_is_done(models.REVIEW_ISSUES_OPEN))
-    self.assertTrue(processes.review_is_done(models.REVIEW_ISSUES_ADDRESSED))
-    self.assertTrue(processes.review_is_done(models.REVIEW_NA))
+    self.assertFalse(processes.review_is_done(core_enums.REVIEW_PENDING))
+    self.assertFalse(processes.review_is_done(core_enums.REVIEW_ISSUES_OPEN))
+    self.assertTrue(processes.review_is_done(core_enums.REVIEW_ISSUES_ADDRESSED))
+    self.assertTrue(processes.review_is_done(core_enums.REVIEW_NA))
 
 
 class ProcessesWellFormedTest(testing_config.CustomTestCase):
@@ -141,9 +142,9 @@ class ProgressDetectorsTest(testing_config.CustomTestCase):
   def setUp(self):
     self.feature_1 = models.Feature(
         name='feature one', summary='sum', category=1, visibility=1,
-        standardization=1, web_dev_views=models.DEV_NO_SIGNALS,
+        standardization=1, web_dev_views=core_enums.DEV_NO_SIGNALS,
         impl_status_chrome=1,
-        intent_stage=models.INTENT_IMPLEMENT)
+        intent_stage=core_enums.INTENT_IMPLEMENT)
     self.feature_1.put()
 
   def tearDown(self):
@@ -164,13 +165,13 @@ class ProgressDetectorsTest(testing_config.CustomTestCase):
   def test_security_review_completed(self):
     detector = processes.PROGRESS_DETECTORS['Security review issues addressed']
     self.assertFalse(detector(self.feature_1))
-    self.feature_1.security_review_status = models.REVIEW_ISSUES_ADDRESSED
+    self.feature_1.security_review_status = core_enums.REVIEW_ISSUES_ADDRESSED
     self.assertTrue(detector(self.feature_1))
 
   def test_privacy_review_completed(self):
     detector = processes.PROGRESS_DETECTORS['Privacy review issues addressed']
     self.assertFalse(detector(self.feature_1))
-    self.feature_1.privacy_review_status = models.REVIEW_ISSUES_ADDRESSED
+    self.feature_1.privacy_review_status = core_enums.REVIEW_ISSUES_ADDRESSED
     self.assertTrue(detector(self.feature_1))
 
   def test_intent_to_prototype_email(self):
@@ -240,19 +241,19 @@ class ProgressDetectorsTest(testing_config.CustomTestCase):
   def test_tag_review_completed(self):
     detector = processes.PROGRESS_DETECTORS['TAG review issues addressed']
     self.assertFalse(detector(self.feature_1))
-    self.feature_1.tag_review_status = models.REVIEW_ISSUES_ADDRESSED
+    self.feature_1.tag_review_status = core_enums.REVIEW_ISSUES_ADDRESSED
     self.assertTrue(detector(self.feature_1))
 
   def test_web_dav_signals(self):
     detector = processes.PROGRESS_DETECTORS['Web developer signals']
     self.assertFalse(detector(self.feature_1))
-    self.feature_1.web_dev_views = models.PUBLIC_SUPPORT
+    self.feature_1.web_dev_views = core_enums.PUBLIC_SUPPORT
     self.assertTrue(detector(self.feature_1))
 
   def test_vendor_signals(self):
     detector = processes.PROGRESS_DETECTORS['Vendor signals']
     self.assertFalse(detector(self.feature_1))
-    self.feature_1.ff_views = models.PUBLIC_SUPPORT
+    self.feature_1.ff_views = core_enums.PUBLIC_SUPPORT
     self.assertTrue(detector(self.feature_1))
 
   def test_estimated_target_milestone(self):
@@ -264,7 +265,7 @@ class ProgressDetectorsTest(testing_config.CustomTestCase):
   def test_code_in_chromium(self):
     detector = processes.PROGRESS_DETECTORS['Code in Chromium']
     self.assertFalse(detector(self.feature_1))
-    self.feature_1.impl_status_chrome = models.ENABLED_BY_DEFAULT
+    self.feature_1.impl_status_chrome = core_enums.ENABLED_BY_DEFAULT
     self.assertTrue(detector(self.feature_1))
 
   def test_motivation(self):
@@ -276,5 +277,5 @@ class ProgressDetectorsTest(testing_config.CustomTestCase):
   def test_code_removed(self):
     detector = processes.PROGRESS_DETECTORS['Code removed']
     self.assertFalse(detector(self.feature_1))
-    self.feature_1.impl_status_chrome = models.REMOVED
+    self.feature_1.impl_status_chrome = core_enums.REMOVED
     self.assertTrue(detector(self.feature_1))
