+++ pages/featuredetail_test.py
@@ -13,8 +13,6 @@
 # limitations under the License.
 
 
-
-
 import testing_config  # Must be imported first
 
 import os
@@ -22,6 +20,7 @@ import flask
 import werkzeug
 import html5lib
 
+from internals import core_enums
 from internals import models
 from framework import ramcache
 from pages import featuredetail
@@ -38,7 +37,7 @@ class TestWithFeature(testing_config.CustomTestCase):
     self.feature_1 = models.Feature(
         name='feature one', summary='detailed sum', category=1, visibility=1,
         standardization=1, web_dev_views=1, impl_status_chrome=1,
-        intent_stage=models.INTENT_IMPLEMENT)
+        intent_stage=core_enums.INTENT_IMPLEMENT)
     self.feature_1.put()
     self.feature_id = self.feature_1.key.integer_id()
